+++ src/main/java/com/example/demo/entity/SysRole.java
@@ -12,7 +12,7 @@ public class SysRole implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue
-    private Long id; // 编号
+    private Integer id; // 编号
     private String role; // 角色标识程序中判断使用,如"admin",这个是唯一的:
     private String description; // 角色描述,UI界面显示使用
     private Boolean available = Boolean.FALSE; // 是否可用,如果不可用将不会添加给用户
@@ -26,17 +26,16 @@ public class SysRole implements Serializable {
     @ManyToMany
     @JoinTable(name="SysUserRole",joinColumns={@JoinColumn(name="roleId")},inverseJoinColumns={@JoinColumn(name="uid")})
     private List<User> userInfos;// 一个角色对应多个用户
-
     public List<User> getUserInfos() {
         return userInfos;
     }
     public void setUserInfos(List<User> userInfos) {
         this.userInfos = userInfos;
     }
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public String getRole() {
